# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookIt SUT API",
    "description": "System Under Test for the BookIt automation framework. All state is namespaced by the X-Test-Session header; X-Test-Now overrides the clock for one request; X-Platform (web, wap, android, ios) enables server-side feature gating.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/studios": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Studios",
        "operationId": "studios_studios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Studio"
                  },
                  "type": "array",
                  "title": "Response Studios Studios Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/schedule": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Schedule",
        "operationId": "schedule_schedule_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Studio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleDay"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/schedule/week": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Schedule Week",
        "operationId": "schedule_week_schedule_week_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleWeek"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/classes/{class_id}": {
      "get": {
        "tags": [
          "catalog"
        ],
        "summary": "Get Class",
        "operationId": "get_class_classes__class_id__get",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Class Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudioClass"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
    "/bookings/{booking_id}/checkin": {
      "get": {
        "tags": [
          "check-in"
        ],
        "summary": "Get Pass",
        "operationId": "get_pass_bookings__booking_id__checkin_get",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Booking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinPass"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/checkins": {
      "post": {
        "tags": [
          "check-in"
        ],
        "summary": "Check In",
        "operationId": "check_in_checkins_post",
        "parameters": [
          {
            "name": "x-studio-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Studio-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResult"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinPass": {
        "properties": {
          "booking_id": {
            "type": "string",
            "title": "Booking Id"
          },
          "class_id": {
            "type": "string",
            "title": "Class Id"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "XXXX-XXXX, static per booking"
          },
          "qr_payload": {
            "type": "string",
            "title": "Qr Payload",
            "description": "bookit:checkin:<code>"
          },
          "window_opens_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Opens At"
          },
          "window_closes_at": {
            "type": "string",
            "format": "date-time",
            "title": "Window Closes At"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_open",
              "open",
              "closed",
              "checked_in"
            ],
            "title": "Status"
          },
          "checked_in_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked In At"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "class_id",
          "code",
          "qr_payload",
          "window_opens_at",
          "window_closes_at",
          "status",
          "checked_in_at"
        ],
        "title": "CheckinPass"
      },
      "CheckinRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "CheckinRequest"
      },
      "CheckinResult": {
        "properties": {
          "booking_id": {
            "type": "string",
            "title": "Booking Id"
          },
          "class_id": {
            "type": "string",
            "title": "Class Id"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "checked_in_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked In At"
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "class_id",
          "class_name",
          "user_name",
          "checked_in_at"
        ],
        "title": "CheckinResult"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorBody"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
...
      "ScheduleDay": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "previous_date": {
            "type": "string",
            "format": "date",
            "title": "Previous Date"
          },
          "next_date": {
            "type": "string",
            "format": "date",
            "title": "Next Date"
          },
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now",
            "description": "The session clock at the time of the request"
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/StudioClass"
            },
            "type": "array",
            "title": "Classes"
          }
        },
        "type": "object",
        "required": [
          "date",
          "previous_date",
          "next_date",
          "now",
          "classes"
        ],
        "title": "ScheduleDay"
      },
      "ScheduleWeek": {
        "properties": {
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start",
            "description": "Monday"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End",
            "description": "Sunday"
          },
          "previous_week": {
            "type": "string",
            "format": "date",
            "title": "Previous Week"
          },
          "next_week": {
            "type": "string",
            "format": "date",
            "title": "Next Week"
          },
          "now": {
            "type": "string",
            "format": "date-time",
            "title": "Now"
          },
          "days": {
            "items": {
              "$ref": "#/components/schemas/ScheduleDay"
            },
            "type": "array",
            "title": "Days"
          }
        },
        "type": "object",
        "required": [
          "week_start",
          "week_end",
          "previous_week",
          "next_week",
          "now",
          "days"
        ],
        "title": "ScheduleWeek"
      },
      "Studio": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "Studio"
      },
      "StudioClass": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "studio_id": {
            "type": "string",
            "title": "Studio Id"
          },
          "studio_name": {
            "type": "string",
            "title": "Studio Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "instructor": {
            "type": "string",
            "title": "Instructor"
          },
          "start_at": {
            "type": "string",
            "format": "date-time",
            "title": "Start At"
          },
          "end_at": {
            "type": "string",
            "format": "date-time",
            "title": "End At"
          },
          "duration_min": {
            "type": "integer",
            "title": "Duration Min"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "spots_left": {
            "type": "integer",
            "title": "Spots Left"
          },
          "is_full": {
            "type": "boolean",
            "title": "Is Full"
          },
          "has_started": {
            "type": "boolean",
            "title": "Has Started"
          },
          "is_outdoor": {
            "type": "boolean",
            "title": "Is Outdoor"
          },
          "is_anchor": {
            "type": "boolean",
            "title": "Is Anchor"
          },
          "my_booking_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Booking Id",
            "description": "The caller's booking on this class, if any"
          },
          "waitlist_count": {
            "type": "integer",
            "title": "Waitlist Count"
          },
          "my_waitlist_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Waitlist Position",
            "description": "The caller's 1-based waitlist position, if any"
          }
        },
        "type": "object",
        "required": [
          "id",
          "studio_id",
          "studio_name",
          "name",
          "category",
          "instructor",
          "start_at",
          "end_at",
          "duration_min",
          "capacity",
          "spots_left",
          "is_full",
          "has_started",
          "is_outdoor",
          "is_anchor",
          "my_booking_id",
          "waitlist_count",
          "my_waitlist_position"
        ],
        "title": "StudioClass"
      },
...
    }
  }
}